requirements.txt
python-telegram-bot[job-queue]==22.0
python-dotenv==1.0.1
aiohttp>=3.9.0
